+++ CarShowroom/src/main/java/com/tuantran/CarShowroom/payload/response/feature/FeatureForFeatureValueResponse.java
@@ -1,12 +1,9 @@
 package com.tuantran.CarShowroom.payload.response.feature;
 
-import com.tuantran.CarShowroom.payload.response.featurevalue.FeatureValueResponse;
 import lombok.Data;
-
-import java.util.Date;
 
 @Data
 public class FeatureForFeatureValueResponse {
-    private int id;
+    private long id;
     private String name;
 }
